Initial refactoring into custom Twitter hosting controller
